remove mandataris-rework page
