+++ app/controllers/mandatarissen/persoon/mandataris.js
@@ -4,28 +4,12 @@ import { action } from '@ember/object';
 import { service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
 
-import { ForkingStore } from '@lblod/ember-submission-form-fields';
-import { SOURCE_GRAPH } from 'frontend-lmb/utils/constants';
-import { getApplicationContextMetaTtl } from 'frontend-lmb/utils/form-context/application-context-meta-ttl';
-
-export default class MandatarissenPersoonMandatarisController extends Controller {
-  @service router;
-  @service store;
-  @service fractieApi;
+export default class MandatarissenPersoonMandatarisReworkController extends Controller {
   @service currentSession;
-  @service('mandataris') mandatarisService;
-
-  @tracked isChanging;
-  @tracked isCorrecting;
-
-  @tracked correctedMandataris = false;
-  @tracked updatedStateMandataris = false;
-  @tracked selectedReplacement = false;
-  @tracked newMandataris;
-
-  @tracked formInitialized;
 
   @tracked isDeleteModalOpen;
+  @tracked isEditModalOpen;
+  @tracked mandatarisFormValues;
 
   get bestuursorganenTitle() {
     const bestuursfunctie = this.model.mandataris.bekleedt
@@ -43,52 +27,6 @@ export default class MandatarissenPersoonMandatarisController extends Controller
     return this.model.mandataris.isBestuurlijkeAliasVan;
   }
 
-  @action
-  async closeModals() {
-    this.isChanging = false;
-    this.isCorrecting = false;
-    this.formInitialized = false;
-    await this.mandatarisService.removeDanglingFractiesInPeriod(
-      this.model.mandataris.id
-    );
-  }
-
-  @action
-  async onUpdateState(newMandataris) {
-    this.editMode = null;
-    this.closeModals();
-    this.newMandataris = newMandataris;
-    this.updatedStateMandataris = true;
-  }
-
-  @action
-  callbackAfterUpdate() {
-    if (this.newMandataris != this.model.mandataris) {
-      this.router.transitionTo(
-        'mandatarissen.mandataris',
-        this.newMandataris.id
-      );
-    }
-  }
-
-  @action
-  async onSave({ instanceTtl }) {
-    this.editMode = null;
-    const store = new ForkingStore();
-    store.parse(instanceTtl, SOURCE_GRAPH, 'text/turtle');
-
-    await this.fractieApi.updateCurrentFractie(this.model.mandataris.id);
-
-    this.correctedMandataris = true;
-    setTimeout(() => this.router.refresh(), 1000);
-    this.closeModals();
-  }
-
-  @action
-  async buildMetaTtl() {
-    return getApplicationContextMetaTtl(this.model.bestuursorganen);
-  }
-
   get isDisabledBecauseLegislatuur() {
     return (
       this.model.periodeHasLegislatuur &&
@@ -99,30 +37,7 @@ export default class MandatarissenPersoonMandatarisController extends Controller
   }
 
   get toolTipText() {
-    if (this.isDisabledBecauseLegislatuur) {
-      return 'Tijdens het voorbereiden van een legislatuur is het niet mogelijk een mandaat in die legislatuur te bewerken.';
-    }
-    if (!this.model.mandataris.isActive) {
-      return 'Het is niet mogelijk de status van een afgelopen mandaat aan te passen.';
-    }
-
-    return 'Dit zal een nieuw mandaat starten met de gewijzigde situatie';
-  }
-
-  get toolTipTextCorrecting() {
-    if (this.isDisabledBecauseLegislatuur) {
-      return 'Tijdens het voorbereiden van een legislatuur is het niet mogelijk een mandaat in die legislatuur te bewerken.';
-    }
-
-    return 'Om fouten te corrigeren, bijvoorbeeld een typefout';
-  }
-
-  get warningTextOCMWLinkToGemeente() {
-    return `Let op! Deze mandataris heeft een gelinkte mandataris in de gemeente.
-      Bij het aanpassen van deze gegevens wordt deze koppeling verbroken.
-      Het doorstromen van gegevens van de gemeente naar OCMW zal
-      hierdoor ook niet meer gebeuren. Om een wijziging aan beide mandaten te
-      maken, gelieve dit te doen in de gemeente.`;
+    return 'Tijdens het voorbereiden van een legislatuur is het niet mogelijk een mandaat in die legislatuur te bewerken.';
   }
 
   get notOwnedByUs() {
@@ -132,5 +47,31 @@ export default class MandatarissenPersoonMandatarisController extends Controller
         (eenheid) => eenheid.id == this.currentSession.group?.id
       )
     );
+  }
+
+  @action
+  async openWizard() {
+    this.mandatarisFormValues = {
+      rangorde: this.model.mandataris.rangorde,
+      start: this.model.mandataris.start,
+      einde: this.model.mandataris.einde,
+      bekleedt: this.model.mandataris.bekleedt,
+      isBestuurlijkeAliasVan: this.model.mandataris.isBestuurlijkeAliasVan,
+      fractie:
+        await this.model.mandataris.heeftLidmaatschap.get('binnenFractie'),
+      status: this.model.mandataris.status,
+      tijdelijkeVervangingen:
+        (await this.model.mandataris.tijdelijkeVervangingen) || [],
+      vervangerVan: (await this.model.mandataris.vervangerVan) || [],
+      beleidsdomein: (await this.model.mandataris.beleidsdomein) || [],
+      contactPoints: (await this.model.mandataris.contactPoints) || [],
+    };
+    this.isEditModalOpen = true;
+  }
+
+  @action
+  closeWizard() {
+    this.mandatarisFormValues = null;
+    this.isEditModalOpen = false;
   }
 }
