+++ app/routes/mandatarissen/persoon/mandataris.js
@@ -3,6 +3,7 @@ import Route from '@ember/routing/route';
 import { service } from '@ember/service';
 
 import { effectiefIsLastPublicationStatus } from 'frontend-lmb/utils/effectief-is-last-publication-status';
+import { isRequiredForBestuursorgaan } from 'frontend-lmb/utils/is-fractie-selector-required';
 import {
   MANDATARIS_EDIT_FORM_ID,
   MANDATARIS_EXTRA_INFO_FORM_ID,
@@ -15,7 +16,6 @@ import RSVP from 'rsvp';
 export default class MandatarissenPersoonMandatarisRoute extends Route {
   @service currentSession;
   @service store;
-  @service router;
   @service semanticFormRepository;
   @service('mandatarissen') mandatarissenService;
 
@@ -28,9 +28,6 @@ export default class MandatarissenPersoonMandatarisRoute extends Route {
     const mandaat = await mandataris.bekleedt;
     const mandatarissen = await this.getMandatarissen(persoon, mandaat);
 
-    const mandatarisEditForm = this.semanticFormRepository.getFormDefinition(
-      MANDATARIS_EDIT_FORM_ID
-    );
     const mandatarisExtraInfoForm =
       await this.semanticFormRepository.getFormDefinition(
         MANDATARIS_EXTRA_INFO_FORM_ID
@@ -71,10 +68,15 @@ export default class MandatarissenPersoonMandatarisRoute extends Route {
         },
       });
     }
+
+    const isFractieRequired = await isRequiredForBestuursorgaan(
+      bestuursorganen.at(0)
+    );
+
     const history = await this.fetchHistory(
       mandataris,
       mandatarissen,
-      mandatarisEditForm.id
+      MANDATARIS_EDIT_FORM_ID
     );
     const isMostRecentVersion =
       [...mandatarissen].sort((a, b) => b.start - a.start)[0].id ===
@@ -84,7 +86,6 @@ export default class MandatarissenPersoonMandatarisRoute extends Route {
       bestuurseenheid,
       mandataris,
       mandatarissen,
-      mandatarisEditForm,
       mandatarisExtraInfoForm,
       history,
       isMostRecentVersion,
@@ -94,6 +95,7 @@ export default class MandatarissenPersoonMandatarisRoute extends Route {
       linkedMandataris,
       owners,
       isDistrictEenheid: isDistrict,
+      isFractieRequired,
       effectiefIsLastPublicationStatus:
         await effectiefIsLastPublicationStatus(mandataris),
       showOCMWLinkedMandatarisWarning,
