--- app/templates/mandatarissen/persoon/mandataris-rework.hbs
@@ -1,113 +1,0 @@
-{{breadcrumb
-  this.model.mandataris.bekleedt.bestuursfunctie.label
-  route="mandatarissen.persoon.mandataris"
-  model=this.model.mandataris.id
-}}
-
-<div class="au-o-grid au-u-padding">
-  <div class="au-o-grid__item au-u-1-2">
-    <p class="au-u-h2 au-u-medium">
-      {{this.bestuursorganenTitle}}
-    </p>
-    <p class="au-u-h3">
-      {{this.persoon.gebruikteVoornaam}}
-      {{this.persoon.achternaam}}
-    </p>
-  </div>
-  <div class="au-o-grid__item au-u-1-2 au-u-text-left au-u-flex--wrap">
-    <div class="au-u-flex au-u-flex--end au-u-margin-bottom">
-      {{#if this.notOwnedByUs}}
-        <AuAlert @skin="info" @icon="eye" @closable={{false}}>
-          <div class="au-u-flex au-u-flex--spaced-tiny">
-            <span>Beheerd door:</span>
-            <Mandaat::OwnershipTags @owners={{this.model.owners}} />
-          </div>
-          <p>Je kan deze mandataris niet bewerken.</p>
-        </AuAlert>
-      {{else}}
-        <AuButtonGroup>
-          <Shared::Tooltip
-            @showTooltip={{this.isDisabledBecauseLegislatuur}}
-            @tooltipText={{this.toolTipText}}
-          >
-            <AuButton
-              @disabled={{this.isDisabledBecauseLegislatuur}}
-              {{on "click" this.openWizard}}
-            >Bewerk mandaat</AuButton>
-          </Shared::Tooltip>
-          <Shared::Tooltip
-            @showTooltip={{not this.model.mandataris.isApprovedForDeletion}}
-            @tooltipText="Mandatarissen die bekrachtigd zijn kunnen niet verwijderd worden."
-          >
-            <AuButton
-              @disabled={{not this.model.mandataris.isApprovedForDeletion}}
-              @skin="secondary"
-              @icon="bin"
-              @alert={{true}}
-              {{on "click" (fn (mut this.isDeleteModalOpen) true)}}
-            >Verwijder</AuButton>
-          </Shared::Tooltip>
-        </AuButtonGroup>
-      {{/if}}
-
-    </div>
-
-  </div>
-</div>
-<div
-  class="au-c-body-container au-c-body-container--scroll au-u-padding-top-none"
->
-  <ValidatieTable
-    @instance={{@model.mandataris}}
-    class="au-u-margin-bottom-none"
-  />
-
-  <div class="au-o-box au-u-padding-top-none">
-    <Mandatarissen::MandatarisCard
-      @mandataris={{@model.mandataris}}
-      @legislatuurInBehandeling={{this.isDisabledBecauseLegislatuur}}
-      @effectiefIsLastPublicationStatus={{this.model.effectiefIsLastPublicationStatus}}
-      @canEdit={{not this.notOwnedByUs}}
-    />
-  </div>
-
-  <div class="au-o-box">
-    <Mandatarissen::MandatarisExtraInfoCard
-      @mandataris={{@model.mandataris}}
-      @form={{@model.mandatarisExtraInfoForm}}
-      @canEdit={{and
-        (not this.isDisabledBecauseLegislatuur)
-        (not this.notOwnedByUs)
-      }}
-    />
-  </div>
-
-  <Mandatarissen::Timeline
-    @mandataris={{@model.mandataris}}
-    @mandatarissen={{@model.mandatarissen}}
-    @history={{@model.history}}
-  />
-
-</div>
-
-<Mandataris::Edit::Wizard
-  @isWizardOpen={{this.isEditModalOpen}}
-  @closeWizard={{this.closeWizard}}
-  @bestuurseenheid={{@model.bestuurseenheid}}
-  @bestuursorgaanIT={{get @model.bestuursorganen 0}}
-  @mandataris={{@model.mandataris}}
-  @mandatarisFormValues={{this.mandatarisFormValues}}
-  @isFractieRequired={{this.model.isFractieRequired}}
-  @showOCMWLinkedMandatarisWarning={{this.showOCMWLinkedMandatarisWarning}}
-/>
-
-<Mandatarissen::DeleteModal
-  @isOpen={{this.isDeleteModalOpen}}
-  @onClose={{fn (mut this.isDeleteModalOpen) false}}
-  @mandataris={{this.model.mandataris}}
-  @linkedMandatarisCanBeDeleted={{and
-    this.model.linkedMandataris.hasDouble
-    this.model.bestuurseenheid.isGemeente
-  }}
-  @linkedMandataris={{this.model.linkedMandataris}}
-/>
